Separate MPI backend. Start MPI backend tests
